+++ test/test.js
@@ -1,10 +1,8 @@
 /*!
  * Copyright (c) 2018-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import * as bedrock from '@bedrock/core';
+import '@bedrock/veres-one-context';
+import '@bedrock/test';
 
-const bedrock = require('bedrock');
-require('bedrock-veres-one-context');
-
-require('bedrock-test');
 bedrock.start();